major changes ch02, ch03, ch04, new ch04mechlogic
Split ch04 on induction into three chapters: doublecheck, then induction, then mechanized logic, with parallel examples in the three chapters. Brutalized ch02. Split out the predicates and quantifiers into their own sections prior to modeling sections. Now predicate logic is treated in a way that is similar to the treatment of propositional logic, and the examples come later. That way, the intro could focus on essentials instead of getting distracted by things like double indexes. Below are a few notes I kept in the revision process to explains why I did some of the things.
 ------------------
I like the Nim model and also the Skolem reasoning, but I think they should be separate because, even though Nim is simpler, there is still a double index to contend with. We introduced propositions without detailed examples,
and I think we can do the same for predicates and quantifiers, so I moved all that stuff into a section before the modelling sections.
 -------------
The Socrates example was a concession to tradition for syllogisms, but I want to limit references to Socrates, Bill likes Mary, Mary likes avocados, and the like to that one case.
 -------------
The forall discussion started with X(3,t), then suddenly switched to X(s,6) for the quantification. The double index is bad enough. Let's not make it worse with bait and switch. Got rid of both in the beginning, and reintroduced the double index later.
 -------------
I think the modeling sections of ch02 still put a higher burden on students than other chapters. The Nim model is a lot easier to follow than tic-tac-toe, but it's still very complicated, and it still has double indexing at its core. That makes it a complicated way to introduce quantifiers. Probably this should be fixed, but I don't think we'll have time to do this before the Jan 1 deadline.
 --------------
I decided to fix it by putting the stuff on quantifier reasoning in new sections before the modeling sections. We introduced Boolean formulas without any special motivation and can do the same for predicates and quantifiers. I think it's better than having nested quantifiers in the first example.
 ---------------
I dont'think we should call the bound variables in quantified formulas "logical variables" because they don't have true/false values. Just call them variables.
 ---------------
Probably a good idea to change the prose proofs in the migration stuff to a sequence of equations, like the proofs we did for Boolean formulas without quantifiers. Leaving this for another pass.
 ---------------
We promised the readers that Boolean formulas would be fully parenthesized (no need to know about operator precedence), except that we might, sometimes, omit the parens at the top level of a formula. I would like to keep that promise, and it's hard to disentangle keep it all straight if we don't do that as we go along. I had to fix the parens in a lot of formulas. It would be a miracle if I got them all right.
 ---------------
\exists x.(P(x)) has redundant parens around P(x). I'd like to avoid redudant parens, too. Fixed a bunch of these. Tried anyway. Plenty of mistakes, no doubt.
 ---------------
I added renaming equations to the table of quantifier equations and
put in a note about restrictions on variable names.
 ---------------
The quanfier equations put the quantified variable as the first argument of the predicate. It doesn't seem right to imply a restriction like this, so I put an ellipsis on both sides of the variable: \forall x.P(... x, ...)
 ---------------
The Nim model in the predicate section used AND (s1 \neq s2) AND ... for the formula that says the stone pile cannot have two different numbers of stones in it, but it should be (s1 \neq s2) IMPLIES ..., I think.
 ---------------
I rather not have more than 2 or 3 exclamation points in the book.
 ---------------
I would like to avoid the term "expression" and, instead, use the term "formula". Expression has too many other meanings in English. Formula is more spacific. Besides it's the essence of "formal".