cleaned up a reference and some algebra
Previously, the formula n(n+1)/2 = 1 + 2 + ... + n was referred to as Gauss's formula. It seems that "triangular number" is a more common name for it, so I changed it to that and added a Wikipedia reference to a proof. Also, I cleaned up some of the algebra in speed-up ratio between isort and msort.