typos and minor revs; Ruben should check
Reverted to axioms for append that require first arg to be a truelist,
as req'd in ACL2 documentation.
Did the same for nthcdr axioms. Left some notes in .tex file.

There was an nthcdr exercise prior to intro of nthcdr. 
Moved it to section where nthcdr is introduced.

Ruben had commented a long time ago that there was something wrong
with the append axioms, or maybe the nthcdr axioms, or maybe both.
Probably a good idea for him to check what's there now.

Spent a long time hacking figure with inductive case of prefix-append 
theorem. It may be a little better, and maybe less fragile.